+++ src/zephyr/tool/environment/get-environments.ts
@@ -1,20 +1,14 @@
 import type { ToolCallback } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { ZodRawShape } from "zod";
+import { Tool } from "../../../common/tools.js";
 import type { ToolParams } from "../../../common/types.js";
-import type { ApiClient } from "../../common/api-client.js";
+import type { ZephyrClient } from "../../client.js";
 import {
   listEnvironmentsQueryParams,
   listEnvironmentsResponse,
 } from "../../common/rest-api-schemas.js";
-import type { ZephyrTool } from "../zephyr-tool.js";
 
-export class GetEnvironments implements ZephyrTool {
-  private readonly apiClient: ApiClient;
-
-  constructor(apiClient: ApiClient) {
-    this.apiClient = apiClient;
-  }
-
+export class GetEnvironments extends Tool<ZephyrClient> {
   specification: ToolParams = {
     title: "Get Environments",
     summary: "Get environments in Zephyr",
@@ -80,10 +74,9 @@ export class GetEnvironments implements ZephyrTool {
 
   handle: ToolCallback<ZodRawShape> = async (args) => {
     const getEnvironmentsInput = listEnvironmentsQueryParams.parse(args);
-    const response = await this.apiClient.get(
-      "/environments",
-      getEnvironmentsInput,
-    );
+    const response = await this.client
+      .getApiClient()
+      .get("/environments", getEnvironmentsInput);
     return {
       structuredContent: response,
       content: [],