+++ src/zephyr/tool/project/get-project.ts
@@ -1,20 +1,14 @@
 import type { ToolCallback } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { ZodRawShape } from "zod";
+import { Tool } from "../../../common/tools.js";
 import type { ToolParams } from "../../../common/types.js";
-import type { ApiClient } from "../../common/api-client.js";
+import type { ZephyrClient } from "../../client.js";
 import {
   getProjectParams,
   getProjectResponse,
 } from "../../common/rest-api-schemas.js";
-import type { ZephyrTool } from "../zephyr-tool.js";
 
-export class GetProject implements ZephyrTool {
-  private readonly apiClient: ApiClient;
-
-  constructor(apiClient: ApiClient) {
-    this.apiClient = apiClient;
-  }
-
+export class GetProject extends Tool<ZephyrClient> {
   specification: ToolParams = {
     title: "Get Project",
     summary: "Get details of project specified by id or key in Zephyr",
@@ -42,7 +36,9 @@ export class GetProject implements ZephyrTool {
 
   handle: ToolCallback<ZodRawShape> = async (args: ZodRawShape) => {
     const { projectIdOrKey } = getProjectParams.parse(args);
-    const response = await this.apiClient.get(`/projects/${projectIdOrKey}`);
+    const response = await this.client
+      .getApiClient()
+      .get(`/projects/${projectIdOrKey}`);
     return {
       structuredContent: response,
       content: [],