+++ src/zephyr/tool/test-cycle/get-test-cycles.ts
@@ -1,20 +1,14 @@
 import type { ToolCallback } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { ZodRawShape } from "zod";
+import { Tool } from "../../../common/tools.js";
 import type { ToolParams } from "../../../common/types.js";
-import type { ApiClient } from "../../common/api-client.js";
+import type { ZephyrClient } from "../../client.js";
 import {
   listTestCyclesQueryParams,
   listTestCyclesResponse,
 } from "../../common/rest-api-schemas.js";
-import type { ZephyrTool } from "../zephyr-tool.js";
 
-export class GetTestCycles implements ZephyrTool {
-  private readonly apiClient: ApiClient;
-
-  constructor(apiClient: ApiClient) {
-    this.apiClient = apiClient;
-  }
-
+export class GetTestCycles extends Tool<ZephyrClient> {
   specification: ToolParams = {
     title: "Get Test Cycles",
     summary: "Get details of Test Cycles in Zephyr",
@@ -77,7 +71,9 @@ export class GetTestCycles implements ZephyrTool {
 
   handle: ToolCallback<ZodRawShape> = async (args) => {
     const parsedArgs = listTestCyclesQueryParams.parse(args);
-    const response = await this.apiClient.get("/testcycles", parsedArgs);
+    const response = await this.client
+      .getApiClient()
+      .get("/testcycles", parsedArgs);
     return {
       structuredContent: response,
       content: [],